+++ compiler/src/test/java/info/teksol/mc/mindcode/SystemLibraryTest.java
@@ -42,6 +42,7 @@ public class SystemLibraryTest {
         return new CompilerProfile(false, OptimizationLevel.ADVANCED)
                 .setProcessorVersion(ProcessorVersion.MAX)
                 .setFinalCodeOutput(FinalCodeOutput.PLAIN)
+                .setAutoPrintflush(false)
                 .setSyntacticMode(SyntacticMode.STRICT)
                 .setOptimizationPasses(25)
                 .setPrintStackTrace(true)